pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hyper-dm"
version = "0.1.0"
description = "Hyper-Diffusion Models for Uncertainty-Aware Medical Image Reconstruction"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "torch>=2.1.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "h5py>=3.9.0",
    "scikit-image>=0.21.0",
    "tqdm>=4.65.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
mri = ["fastmri>=0.3.0"]
dev = ["pytest>=7.4.0", "ruff>=0.1.0"]
mlops = ["mlflow>=2.9.0"]

[tool.setuptools.packages.find]
include = ["hyper_dm*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
